:rotating_light: fix clippy warnings
