--- marketplace-indexer/src/domain/observers/event_store/mod.rs
@@ -1,5 +1,0 @@
-mod event_store;
-pub use event_store::*;
-
-#[cfg(test)]
-mod test;
